Add files via upload
Had an member variable in the AWindowsLib Status Bar Item class that when the status bar was constructed would attempt to set a string based on the variable's unpredictable value sometime causing lockups at start up.

+++ Dialogs.cpp
@@ -2019,7 +2019,7 @@ void HashTagChooseDlg::OnListViewItemChanged(int child, NMLISTVIEW *pLVN)
  if (child != IDC_HASHTAG_CHOOSE_LIST) 
    return;
 
- id = pLVN->lParam;
+ id = (int)pLVN->lParam;
 
  switch(m_HashTagType)
   {
